Handle the notification Able to cancel the run Save the run to the database
Problem: Cannot draw the polyline on the screen

+++ app/src/main/java/com/example/runningtrackingapp/db/Run.kt
@@ -13,7 +13,7 @@ data class Run (
     var avgSpeedInKMH:Float = 0f,
     var distanceInMater :Float = 0f,
     var timeInMillis:Long = 0L, //Determines the time of each run
-    var caloriesBurnt :Float = 0f)
+    var caloriesBurnt :Int = 0)
 {
     @PrimaryKey(autoGenerate = true)
     var id :Int? = null
